move everything to package cmd so everything can access the helpers
